requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
celery[redis]==5.5.2
redis>=5.0.0
pillow==11.2.1
python-multipart==0.0.20
flower==2.0.1
